+++ pkg/kubectl/cmd/rollingupdate.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/api"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/kubectl"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/kubectl/cmd/util"
 	"github.com/spf13/cobra"
 )
 
@@ -48,13 +49,13 @@ $ kubectl rollingupdate frontend-v1 -f frontend-v2.json
 $ cat frontend-v2.json | kubectl rollingupdate frontend-v1 -f -
   <update pods of frontend-v1 using json data passed into stdin>`,
 		Run: func(cmd *cobra.Command, args []string) {
-			filename := GetFlagString(cmd, "filename")
+			filename := util.GetFlagString(cmd, "filename")
 			if len(filename) == 0 {
 				usageError(cmd, "Must specify filename for new controller")
 			}
-			period := GetFlagDuration(cmd, "update-period")
-			interval := GetFlagDuration(cmd, "poll-interval")
-			timeout := GetFlagDuration(cmd, "timeout")
+			period := util.GetFlagDuration(cmd, "update-period")
+			interval := util.GetFlagDuration(cmd, "poll-interval")
+			timeout := util.GetFlagDuration(cmd, "timeout")
 			if len(args) != 1 {
 				usageError(cmd, "Must specify the controller to update")
 			}
@@ -67,7 +68,7 @@ $ cat frontend-v2.json | kubectl rollingupdate frontend-v1 -f -
 			cmdApiVersion := clientConfig.Version
 
 			mapper, typer := f.Object(cmd)
-			mapping, namespace, newName, data := ResourceFromFile(filename, typer, mapper, schema, cmdApiVersion)
+			mapping, namespace, newName, data := util.ResourceFromFile(filename, typer, mapper, schema, cmdApiVersion)
 			if mapping.Kind != "ReplicationController" {
 				usageError(cmd, "%s does not specify a valid ReplicationController", filename)
 			}
@@ -78,7 +79,7 @@ $ cat frontend-v2.json | kubectl rollingupdate frontend-v1 -f -
 
 			cmdNamespace, err := f.DefaultNamespace(cmd)
 			checkErr(err)
-			err = CompareNamespace(cmdNamespace, namespace)
+			err = util.CompareNamespace(cmdNamespace, namespace)
 			checkErr(err)
 
 			client, err := f.Client(cmd)
